fix: Increase default export timeout to 30 minutes
It often happens in the Bitcoin blockchain that the interval between
two consecutive blocks is much higher than 10 minutes and also higher
than 15. The previous timeout leads to many unnecessary restarts
because a healthcheck fails whenever no new block has appeared.

We increase the default timeout to 30 minutes to reduce the number of
restarts.